Added platform compile guards. Moved PageTableManager to arch-specific folder. Added platform-specific defines file, automatically included in platform.h.
