+++ kernel/PageFrameAllocator.cpp
@@ -1,5 +1,8 @@
 #include <StringExtras.h>
 #include <PageFrameAllocator.h>
+#include <Platform.h>
+
+PLATFORM_REQUIRED(MINOS_PLATFORM_X86_64)
 
 namespace Kernel
 {
