+++ kernel/include/arch/x86_64/ArchitectureDefs.h
@@ -1,0 +1,10 @@
+#pragma once
+
+/*
+    This file contains any platform specific magic numbers. I'd rather keep them in one place, rather than
+    all over the code base.
+*/
+
+#define PAGE_SIZE 0x1000
+
+#define X86_MSR_EFER 0xC0000080
